+++ src/cc/view.rs
@@ -526,9 +526,9 @@ impl View {
         // Remove from suffstats, unassign, and drop components if singleton.
         // Get a list of the components that were removed so we can update the
         // assignment to preserve canonical order.
-        (ix..ix + n).for_each(|row_ix| {
-            self.remove_row(row_ix);
-            self.asgn.asgn.remove(row_ix);
+        (ix..ix + n).for_each(|_| {
+            self.remove_row(ix);
+            self.asgn.asgn.remove(ix);
         });
 
         // remove data from features